feat: Redis 관련 API 추가 및 적용

+++ src/main/java/jungle/HandTris/domain/GameRoom.java
@@ -20,9 +20,9 @@ public class GameRoom {
     @Enumerated(value = EnumType.STRING)
     private GameCategory gameCategory;
 
-    private long participantCount;
+    private int participantCount;
 
-    private long participantLimit;
+    private int participantLimit;
 
     @Column(nullable = false)
     private UUID uuid;
